Add custom primaries and illuminant choice to ICCv2 - solve some bugs

+++ rtgui/iccprofilecreator.h
@@ -95,7 +95,7 @@ private:
     void adjusterAutoToggled(Adjuster* a, bool newval);
     static std::vector<Glib::ustring> getGamma();
     Glib::ustring getPrimariesPresetName(const Glib::ustring &preset);
-    void getPrimaries(const Glib::ustring &preset, float *p, ColorTemp &temp);
+    void getPrimaries(const Glib::ustring &preset, double *p, ColorTemp &temp);
     Glib::ustring getGammaPresetName(const Glib::ustring &preset);
     void getGamma(const Glib::ustring &preset, double &gamma, double &slope);
     void savePressed();
